+++ crates/ev-zkevm-types/src/programs/hyperlane/mod.rs
@@ -1,5 +1,6 @@
 pub mod tree;
 pub mod types;
+
 use sha3::{Digest, Keccak256};
 
 pub fn digest_keccak(bytes: &[u8]) -> [u8; 32] {
