+++ crates/ev-zkevm-types/src/programs/hyperlane/tree.rs
@@ -136,10 +136,10 @@ impl MerkleTree {
 
 #[cfg(test)]
 mod tests {
-    use crate::programs::hyperlane::tree::{MerkleTree, ZERO_HASHES};
     use alloy_primitives::Keccak256;
     use anyhow::Result;
 
+    use crate::programs::hyperlane::tree::{MerkleTree, ZERO_HASHES};
     #[test]
     fn test_insert() {
         let mut tree = MerkleTree::default();
