+++ crates/storage/src/hyperlane/message.rs
@@ -2,10 +2,11 @@
 /// It is used to store and retrieve Hyperlane messages.
 /// The messages are stored in a column family called "messages".
 use anyhow::{Context, Result};
-use ev_zkevm_types::StoredHyperlaneMessage;
 use rocksdb::{ColumnFamilyDescriptor, DB, IteratorMode, Options};
 use std::path::Path;
 use std::sync::{Arc, RwLock};
+
+use crate::hyperlane::StoredHyperlaneMessage;
 
 pub struct HyperlaneMessageStore {
     pub db: Arc<RwLock<DB>>,
