+++ crates/storage/src/hyperlane/mod.rs
@@ -1,15 +1,28 @@
+use ev_zkevm_types::hyperlane::HyperlaneMessage;
+use serde::{Deserialize, Serialize};
+
 pub mod message;
 pub mod snapshot;
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+pub struct StoredHyperlaneMessage {
+    pub block_number: Option<u64>,
+    pub message: HyperlaneMessage,
+}
+
+impl StoredHyperlaneMessage {
+    pub fn new(message: HyperlaneMessage, block_number: Option<u64>) -> Self {
+        Self { block_number, message }
+    }
+}
 
 #[cfg(test)]
 mod tests {
     use crate::{
         APP_HOME,
-        hyperlane::{message::HyperlaneMessageStore, snapshot::HyperlaneSnapshotStore},
+        hyperlane::{StoredHyperlaneMessage, message::HyperlaneMessageStore, snapshot::HyperlaneSnapshotStore},
     };
-    use ev_zkevm_types::{
-        StoredHyperlaneMessage, hyperlane::decode_hyperlane_message, programs::hyperlane::tree::MerkleTree,
-    };
+    use ev_zkevm_types::{hyperlane::decode_hyperlane_message, programs::hyperlane::tree::MerkleTree};
 
     const DEFAULT_MESSAGE: &str = "0300000009000004d2000000000000000000000000a7578551bae89a96c3365b93493ad2d4ebcbae9700010f2c726f757465725f617070000000000000000000000000000100000000000000000000000000000000000000006a809b36caf0d46a935ee76835065ec5a8b3cea700000000000000000000000000000000000000000000000000000000000003e8";
